+++ piGantry/piSerial/motorFunc.py
@@ -3,7 +3,7 @@ from piGantry.piSerial import mathFunc
 import time
 
 class motor():
-    def __init__(self):
+    def __init__(self, gearRatio):
         self.zeroDone = False
         self.startZero = False
         self.moveReady = False
@@ -15,6 +15,7 @@ class motor():
         self.startOffset = int(0)
         self.endOffset = int(0)
         self.pulsePerRev = int(12800)
+        self.gearRatio = float(gearRatio)
 
 # This takes the digital dial reading & produces # of steps 
 # for the clearCore during zeroing 
@@ -38,12 +39,12 @@ def formatMsg(dialRead, target):
 # serialDevices should be tuple of 3 devices, (clearCore, micro, laser)
 
 # Handles zeroing process by reading digital dial & issuing instructions to clearCore
-def runZero(motor, serialDevices, microZero=True):
+def runZero(motor, serialDevices, zeroPoint, microZero=True):
     if (type(serialDevices) is tuple):
         clearCore = serialDevices[0]
     else:
         clearCore = serialDevices
-    while not motor.startZero :
+    while not motor.startZero:
         if (clearCore.readIn() == "start"):
             motor.startZero = True
             clearCore.writeOut("start")
@@ -61,7 +62,7 @@ def runZero(motor, serialDevices, microZero=True):
                     else:
                         # Call on readDial function passing micro.port initialized in class constructor
                         input = serialComm.readDial(serialDevices[1].port)
-                        out = formatMsg(input, 4)
+                        out = formatMsg(input, zeroPoint)
                         clearCore.writeOut(out)
                         print(out)
                         if (out == "stp"):
@@ -79,11 +80,13 @@ def runZero(motor, serialDevices, microZero=True):
 
 def runOneMove(motor, clearCore, stepsAdjusted):
     while (not motor.moveReady):
+        print("Waiting for move")
         if (clearCore.readIn() == "move"):
-            motor.moveReady = True
+            motor.moveReady = True 
             motor.moveDone = False
+        print("Hello!!")
     if (motor.moveReady):
-        time.sleep(2)
+        time.sleep(0.1)
         clearCore.writeOut("move")
         clearCore.writeOut(f"{stepsAdjusted}")
         motor.moveReady = False
@@ -92,19 +95,34 @@ def runOneMove(motor, clearCore, stepsAdjusted):
             print("move done")
             motor.moveDone = True
 
-def runMoves(steps1, motorObj, serialDevices, steps2 = False, straightHome = True):
+def adjustToEncoder(clearCore, motor, encoder, meterDistance):
+    encoderReading = serialComm.readArduinoEncoder(encoder)
+    print(f"encoder reading {encoderReading}")
+    print(f"meter dist 1 {meterDistance - 0.0001}")
+    print(f"meter dist 2 {meterDistance + 0.0001}")
+    while ((meterDistance - encoderReading) > 0.1 or (meterDistance - encoderReading) > -0.1):
+        encoderReading = serialComm.readArduinoEncoder(encoder)
+        diff = (meterDistance - encoderReading)
+        stepsToTravel = mathFunc.calcDist(19200, diff, convertMMToSteps=True)
+        if stepsToTravel < 1:
+            break
+        stepsToTravel = round(stepsToTravel)
+        print(f"steps to travel {stepsToTravel} | diff {diff} | encoder reading {encoderReading}")
+        runOneMove(motor, clearCore, stepsToTravel)
+        time.sleep(0.3)
+
+def runMoves(steps1, motorObj, serialDevices, steps2 = False, straightHome = True, encoder = False, jiggle = False, compensation = True):
     # amountOfSteps+1 for returning back to zero in one move, 
     # amountOfSteps*2 for returning back to zero in same amount of moves & steps per move
     print(f"{type(motorObj)} yo")
+    wiggleHome = False
     if (type(serialDevices) is tuple):
         clearCore = serialDevices[0]
         clearCore2 = serialDevices[1]
     else:
         clearCore = serialDevices
     if (type(motorObj) is tuple):
-        print(f"{type(motorObj)} yo2")
         motor = motorObj[0]
-        print(f"{type(motorObj)} yo3")
         motor2 = motorObj[1]
     else:
         motor = motorObj
@@ -118,17 +136,68 @@ def runMoves(steps1, motorObj, serialDevices, steps2 = False, straightHome = Tru
         stepsRange = amountOfSteps*2
         stepMulti = -1
 
+    if (type(encoder) is tuple):
+        encoderReader = encoder[1]
+
+    correctionVal = 0
+    totalCorrectionVal = 0
     for i in range(stepsRange):
+        print(f"\n\nMove number: {i}\n")
+        print("Move start")
         stepsAdjusted = steps
+        print(f"Correction value: {correctionVal}")
         if (i >= amountOfSteps):
-            stepsAdjusted = (steps*stepMulti)
+            stepsAdjusted = (steps*stepMulti)-totalCorrectionVal
+            #if not wiggleHome:
+            #    for i in range(2):
+            #        runOneMove(motor, clearCore, 200)
+            #        runOneMove(motor, clearCore, -200)
+            #wiggleHome = True
+            print("Going home")
+            print(serialComm.readArduinoEncoder(encoderReader))
+            foo = input("Press Enter to continue...")
+            #time.sleep(15)
+        elif compensation:
+            stepsAdjusted = stepsAdjusted + correctionVal
+        correctionVal = 0
 
-        mmDistance = mathFunc.calcDist(12800, stepsAdjusted)
+        mmDistance = mathFunc.calcDist(12800, stepsAdjusted, motor.gearRatio)
         meterDistance = float(mmDistance / 1000)
-        totalStep = float(mathFunc.calcDist(12800, steps * (i + 1))/1000)
+        totalStep = float(mathFunc.calcDist(12800, steps * (i + 1), motor.gearRatio)/1000)
 
-        print(f"Total calcualted step distance = {totalStep}m")
+        print(f"Total calcualted step distance 1 = {totalStep}m")
         if steps2:
             runMoves(steps2, motor2, clearCore2)
-        runOneMove(motor, clearCore, stepsAdjusted)
+        runOneMove(motor, clearCore, (stepsAdjusted))
+        if (not i >= amountOfSteps):
+            if encoder:
+                reading = serialComm.readArduinoEncoder(encoder[1])
+                if encoder[0]:
+                    time.sleep(5)
+                    print(f"Encoder read 1: {reading}")
+                    adjustToEncoder(clearCore, motor, encoder[1], meterDistance)
+                    print("Adjust finished!")
+                diff = (totalStep - reading)*1000
+                print(f"Diff value: {diff}")
+                print(f"Calculated dist: {totalStep} | Encoder read: {reading}")
+                stepsToTravel = mathFunc.calcDist(19200, diff, motor.gearRatio, convertMMToSteps=True)
+                stepsToTravel = round(stepsToTravel)
+                if compensation:
+                    if stepsToTravel > 1 or stepsToTravel < -1:
+                        correctionVal = stepsToTravel
+                    else:
+                        correctionVal = 0
+                    totalCorrectionVal += correctionVal
+                print(f"Steps value: {stepsToTravel} | Correction Value: {correctionVal} | Total correction value: {totalCorrectionVal}")
+                
+
+        if jiggle:
+            for i in range(2):
+                runOneMove(motor, clearCore, 200)
+                runOneMove(motor, clearCore, -200)
+        if encoder[0]:
+            print(f"Encoder read 2: {serialComm.readArduinoEncoder(encoder[1])}")
+        #print("Take reading!")
+        #time.sleep(15)
+        print("Move end")
     return True