+++ piGantry/piSerial/serialInter.py
@@ -2,12 +2,12 @@ import serial
 import time
 
 class serialObject(object):
-    def __init__(self, baudRate, serialPortName, timeout=1, writeTimeOut= 10):
+    def __init__(self, baudRate, serialPortName, timeout=1, writeTimeOut= 3):
         self.newDataOut = False
         self.newDataIn = False
         self.confirmMessage = False
         # Initialize serial object
-        self.port = serial.Serial(serialPortName, baudrate = baudRate, timeout= 1,  write_timeout=10, rtscts = False)
+        self.port = serial.Serial(serialPortName, baudrate = baudRate, timeout= timeout,  write_timeout=writeTimeOut, rtscts = False)
 
     def writeOut(self, msg):
         try:
@@ -18,7 +18,7 @@ class serialObject(object):
         except:
             print("Write failed, trying again")
 
-    def readIn(self):
+    def readIn(self, sendMsg = False):
         msgInString = False
         while not msgInString:
             if (self.port.inWaiting() > 0):
@@ -29,6 +29,52 @@ class serialObject(object):
                 return msgInString
             else:
                 self.newDataIn = False
+                if sendMsg:
+                    print(f"No input, sending msg : {sendMsg}")
+                    self.writeOut(sendMsg)
+                    time.sleep(0.01)
+
+def initializeArduinoEncoder(encoder):
+    counter = 0
+    while True:
+        encoder.writeOut("zero")
+        if (encoder.readIn() == "start"):
+            encoder.writeOut("start")
+        else:
+            counter += 1
+            time.sleep(0.1)
+            if counter > 5:
+                break
+
+def zeroArduinoEncoder(encoder):
+    zeroDone = False
+    print("hello")
+    while not zeroDone:
+        encoder.writeOut("zero")
+        time.sleep(0.2)
+        print("hello1")
+        msg = encoder.readIn(sendMsg = "zero")
+        print("hello3")
+        print(msg)
+        if (msg == "zero"):
+            zeroDone = True
+            break
+        print("hello2")
+
+def readArduinoEncoder(encoder):
+    encoder.writeOut("read")
+    valEncoder = False
+    encoder.port.flushInput()
+    encoder.port.flushOutput()
+    while not valEncoder:
+        print(encoder.port.inWaiting())
+        if encoder.port.inWaiting() > 0:
+            valEncoder = encoder.port.readline().decode().strip()
+            return (float(valEncoder)*5)*1e-6
+        else:
+            encoder.writeOut("read")
+        time.sleep(0.5)
+
 
 # Will collapse readDial & readLaser w/ DRY in mind   
 def readDial(port):